% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opcc.R
\name{geonames_supplementary_adapter}
\alias{geonames_supplementary_adapter}
\title{Load the versioned GeoNames supplementary-point adapter}
\usage{
geonames_supplementary_adapter()
}
\value{
An \code{opcc_source_adapter} for the packaged GeoNames point artifact.
}
\description{
Load the versioned GeoNames supplementary-point adapter
}
