% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opcc.R
\name{list_vintages}
\alias{list_vintages}
\title{List supported correspondence release vintages}
\usage{
list_vintages(level = c("DB", "DA"))
}
\arguments{
\item{level}{Geography level, \code{"DB"} or \code{"DA"}.}
}
\value{
A character vector of release vintages.
}
\description{
List supported correspondence release vintages
}
\examples{
list_vintages("DB")
list_vintages("DA")

}
