% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opcc.R
\name{aggregate_da_correspondence}
\alias{aggregate_da_correspondence}
\title{Aggregate postal-code-to-DB evidence to DA links}
\usage{
aggregate_da_correspondence(correspondence)
}
\arguments{
\item{correspondence}{A postal-code-to-DB correspondence data frame.}
}
\value{
A data frame with one row per \code{postal_code} and \code{DAUID}.
}
\description{
Aggregate postal-code-to-DB evidence to DA links
}
